fix: suppress PHP 7.x type errors
The `mixed` psuedo-type is not supported in PHP 7.x, which results
in fatal uncaught TypeErrors. Using the ReturnTypeWillChange attribute
to temporarily suppress return type mismatch errors allows us to
continue supporting PHP 7.x for the time being.

The return type of `jsonSerializable` will need to be updated to
`mixed` before PHP 9 to avoid fatal errors, which will require
bumping the minimum PHP version for `php-taxcloud` and dependants
to at least 8.0.